requirements.txt
absl-py==2.2.1
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.1
anyio==4.8.0
appnope==0.1.4
arfpy==0.1.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autograd==1.7.0
autograd-gamma==0.5.0
babel==2.17.0
be_great==0.0.8
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
Boruta==0.4.3
boto3==1.37.23
botocore==1.37.23
Brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
catboost==1.2.7
category_encoders==2.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
copulas==0.12.1
ctgan==0.11.0
cycler==0.12.1
cymem==2.0.11
databricks-sdk==0.49.0
datasets==3.5.0
debugpy==1.8.13
decaf-synthetic-data==0.1.6
decorator==5.2.1
deepecho==0.7.0
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
docker==7.1.0
executing==2.2.0
Faker==37.1.0
fastai==2.7.19
fastapi==0.115.12
fastcore==1.7.29
fastdownload==0.0.7
fastjsonschema==2.21.1
fastprogress==1.0.3
feather-format==0.4.1
fflows==0.0.3
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
formulaic==1.1.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
geomloss==0.2.6
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.0
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
inflate64==1.0.1
interface-meta==1.3.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keopscore==2.2.3
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lifelines==0.29.0
lightning-utilities==0.14.2
llvmlite==0.44.0
loguru==0.7.3
Mako==1.3.9
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mlflow==2.21.2
mlflow-skinny==2.21.2
monai==1.4.0
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
multivolumefile==0.2.3
murmurhash==1.0.12
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
nflows==0.14
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
opacus==1.5.3
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opt_einsum==3.4.0
optuna==4.2.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pgmpy==0.1.26
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.1
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==0.22.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.3
pybind11==2.13.6
pycox==0.3.0
pycparser==2.22
pycryptodomex==3.22.0
pydantic==1.10.21
Pygments==2.19.1
pykeops==2.2.3
pyparsing==3.2.1
pyppmd==1.1.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytorch-lightning==1.9.5
pyts==0.13.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
pyzstd==0.16.2
rdt==1.15.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sdmetrics==0.19.0
sdv==1.19.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==76.0.0
shap==0.47.1
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.39
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
sympy==1.13.1
synthcity==0.2.11
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminado==0.18.1
texttable==1.7.0
thinc==8.3.4
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.2.2
torchmetrics==1.7.0
torchtext==0.17.2
torchtuples==0.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
tsai==0.4.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xgboost==2.1.4
xgbse==0.3.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
